Fix power overflow issues when checking for a potential strong quorum (#590)
* Use int64 for power calculations, not uint16

Fixes an overflow when computing the maximum amount of support a
proposal could have had. There's no reason to use uint16 here anyways,
even if power values are guaranteed to fit.

* Limit power to the max available in `CouldReachStrongQuorumFor`

We're intentionally double-counting adversary power here, so should
handle the case where our power sums to over the total.

+++ gpbft/ticket_quality.go
@@ -19,7 +19,7 @@ import (
 // This ends up being `-log(ticket) / power` where ticket is [0, 1).
 // We additionally use log-base-2 instead of natural logarithm as it is easier to implement,
 // and it is just a linear factor on all tickets, meaning it does not influence their ordering.
-func ComputeTicketQuality(ticket []byte, power uint16) float64 {
+func ComputeTicketQuality(ticket []byte, power int64) float64 {
 	// we could use Blake2b-128 but 256 is more common and more widely supported
 	ticketHash := blake2b.Sum256(ticket)
 	quality := linearToExpDist(ticketHash[:16])
